refactor: fetch-with-timeout, nonce CSP, chat race, server-component pages
- lib/fetch-with-timeout.ts: hardened fetch wrapper with AbortSignal.timeout
  default 10s, exponential backoff + jitter, retry on 429/502/503, and
  Retry-After honoring. Rolled out to weather/current and
  space-weather/kp-index as the starter set.
- lib/api/error-response.ts: apiError(code, message, opts) helper with a
  typed ApiErrorCode union (BAD_REQUEST, UNAUTHORIZED, FORBIDDEN,
  NOT_FOUND, RATE_LIMITED, UPSTREAM_ERROR, INTERNAL_ERROR). Applied to
  weather/current upstream failures.
- proxy.ts: generate a per-request CSP nonce, set on x-nonce request
  header (Next.js stamps it onto hydration scripts), and emit the CSP
  response header. Prod uses `script-src 'self' 'nonce-<n>'
  'strict-dynamic'` — no more 'unsafe-inline'. Non-prod keeps the looser
  policy for the Vercel overlay. Static CSP removed from next.config.mjs
  to avoid collision with the dynamic one.
- lib/services/chat-rate-limiter.ts: optimistic-lock increment closes the
  TOCTOU race where two concurrent requests could both see count=29 and
  both increment. On lost updates, re-read and only allow if still under
  the limit. app/api/chat/route.ts now emits Retry-After and
  X-RateLimit-* headers on 429.
- De-client page shells so they can export static metadata and SSG:
  tropical (pure static), severe/situation (alerts leaf as client
  component), hourly (client body in leaf, page is server w/
  force-dynamic for searchParams).
