+++ next.config.mjs
@@ -47,9 +47,11 @@ const nextConfig = {
   // Optimize for production
   productionBrowserSourceMaps: false,
 
-  // Add headers for better caching and security
+  // Add headers for better caching and security.
+  // NOTE: Content-Security-Policy is set per-request in proxy.ts so we can
+  // use a fresh nonce for script-src in production. Do NOT duplicate CSP
+  // here — a static CSP would collide with the nonce CSP.
   async headers() {
-    const isProd = process.env.NODE_ENV === 'production'
     return [
       {
         source: '/:path*',
@@ -78,26 +80,6 @@ const nextConfig = {
           {
             key: 'Referrer-Policy',
             value: 'strict-origin-when-cross-origin'
-          },
-          {
-            key: 'Content-Security-Policy',
-            value: [
-              "default-src 'self'",
-              "script-src 'self' 'unsafe-inline' https://vercel.live https://vercel.com https://va.vercel-scripts.com",
-              "style-src 'self' 'unsafe-inline' https://fonts.googleapis.com",
-              "font-src 'self' https://fonts.gstatic.com data:",
-              "img-src 'self' data: https: blob:",
-              "connect-src 'self' https://api.openweathermap.org https://pollen.googleapis.com https://www.google.com https://*.supabase.co https://*.sentry.io https://vitals.vercel-insights.com https://mesonet.agron.iastate.edu https://tile.openstreetmap.org",
-              "worker-src 'self' blob:",
-              "frame-src 'self'",
-              "object-src 'none'",
-              "base-uri 'self'",
-              "form-action 'self'",
-              "frame-ancestors 'self'",
-              // Only enable upgrade-insecure-requests in production; it breaks local http dev
-              // by upgrading same-origin asset requests (notably in WebKit).
-              ...(isProd ? ["upgrade-insecure-requests"] : [])
-            ].join('; ')
           },
           {
             key: 'Permissions-Policy',
